feat: make xp boost scrollable in PlayerStatsBar
